requirements.txt
asgiref==3.4.1
boto3==1.20.45
botocore==1.23.45
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.11.0
django-crispy-forms==1.13.0
django-extensions==3.1.5
django-jazzmin==2.4.8
graphviz==0.19.1
jmespath==0.10.0
Pillow==8.4.0
psycopg2-binary==2.9.3
pygraphviz==1.8
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
s3transfer==0.5.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.8
